encode decode e2e
It can encode and decode end 2 end, but does not pass the "cmp" test. Which means it does not decompressed back to the original content. The format is messed. Original content (sherlock) uses UTF8, I am trying to make it use ASCII. May need to figure out some other encoder for literals.